move streams to traits package
