+++ src/main/java/com/aol/simple/react/stream/eager/EagerReact.java
@@ -21,6 +21,7 @@ import com.aol.simple.react.generators.Generator;
 import com.aol.simple.react.generators.ReactIterator;
 import com.aol.simple.react.stream.BaseSimpleReact;
 import com.aol.simple.react.stream.ThreadPools;
+import com.aol.simple.react.stream.traits.EagerFutureStream;
 import com.nurkiewicz.asyncretry.RetryExecutor;
 
 /**
