+++ src/test/java/com/aol/simple/react/collectors/lazy/BatchingCollectorTest.java
@@ -14,8 +14,8 @@ import org.junit.Test;
 import org.mockito.Mockito;
 
 import com.aol.simple.react.config.MaxActive;
-import com.aol.simple.react.stream.eager.EagerFutureStream;
-import com.aol.simple.react.stream.lazy.LazyFutureStream;
+import com.aol.simple.react.stream.traits.EagerFutureStream;
+import com.aol.simple.react.stream.traits.LazyFutureStream;
 
 public class BatchingCollectorTest {
