+++ src/test/java/com/aol/simple/react/eager/AsyncDisplayTest.java
@@ -4,8 +4,8 @@ import java.util.stream.Stream;
 
 import org.junit.Test;
 
-import com.aol.simple.react.stream.eager.EagerFutureStream;
-import com.aol.simple.react.stream.lazy.LazyFutureStream;
+import com.aol.simple.react.stream.traits.EagerFutureStream;
+import com.aol.simple.react.stream.traits.LazyFutureStream;
 
 public class AsyncDisplayTest {
