+++ src/test/java/com/aol/simple/react/eager/JDKStreamTest.java
@@ -2,8 +2,8 @@ package com.aol.simple.react.eager;
 
 import java.util.stream.Stream;
 
-import com.aol.simple.react.stream.eager.EagerFutureStream;
-import com.aol.simple.react.stream.eager.EagerFutureStream.*;
+import com.aol.simple.react.stream.traits.EagerFutureStream;
+import com.aol.simple.react.stream.traits.EagerFutureStream.*;
 import com.aol.simple.react.base.BaseJDKStreamTest;
 
 public class JDKStreamTest extends BaseJDKStreamTest{
