+++ src/test/java/com/aol/simple/react/lazy/LazySeqTest.java
@@ -1,12 +1,9 @@
 package com.aol.simple.react.lazy;
 
-import static com.aol.simple.react.stream.lazy.LazyFutureStream.parallel;
-import static com.aol.simple.react.stream.lazy.LazyFutureStream.parallelBuilder;
+import static com.aol.simple.react.stream.traits.LazyFutureStream.parallel;
+import static com.aol.simple.react.stream.traits.LazyFutureStream.parallelBuilder;
 import static java.util.Arrays.asList;
-import static org.hamcrest.Matchers.containsInAnyOrder;
-import static org.hamcrest.Matchers.equalTo;
 import static org.hamcrest.Matchers.*;
-import static org.hamcrest.Matchers.lessThan;
 import static org.jooq.lambda.tuple.Tuple.tuple;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertThat;
@@ -29,10 +26,10 @@ import com.aol.simple.react.async.Queue;
 import com.aol.simple.react.async.QueueFactories;
 import com.aol.simple.react.base.BaseSeqTest;
 import com.aol.simple.react.stream.ThreadPools;
-import com.aol.simple.react.stream.eager.EagerFutureStream;
-import com.aol.simple.react.stream.lazy.LazyFutureStream;
 import com.aol.simple.react.stream.simple.SimpleReact;
+import com.aol.simple.react.stream.traits.EagerFutureStream;
 import com.aol.simple.react.stream.traits.FutureStream;
+import com.aol.simple.react.stream.traits.LazyFutureStream;
 
 public class LazySeqTest extends BaseSeqTest {
