+++ src/test/java/com/aol/simple/react/simple/MergeTest.java
@@ -11,8 +11,8 @@ import java.util.concurrent.ExecutionException;
 import org.junit.Test;
 
 import com.aol.simple.react.stream.ThreadPools;
-import com.aol.simple.react.stream.eager.EagerFutureStream;
 import com.aol.simple.react.stream.simple.SimpleReact;
+import com.aol.simple.react.stream.traits.EagerFutureStream;
 import com.aol.simple.react.stream.traits.SimpleReactStream;
 
 public class MergeTest {
